Used vuetify components in add or edit user component and addressed some of the review comments
