+++ vue.config.js
@@ -1,6 +1,5 @@
 module.exports = {
   transpileDependencies: ["vuetify"],
-  lintOnSave: false,
   pluginOptions: {
     i18n: {
       locale: "en",
